+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog],
 and this project adheres to [Semantic Versioning].
 
+## [0.6.0] - 2022-05-16
+### Added
+- `seq get` command for retrieval of FASTQ/FASTA sequences.
+
+### Changed
+- `tree add` command now uses Analysis abbreviation.
+- `metadata add` command now uses Species abbreviation.
+
 ## [0.5.3] - 2022-05-10
 ### Fixed
 - Fixed broken build.
@@ -117,3 +125,4 @@ and this project adheres to [Semantic Versioning].
 [0.5.1]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.5.1
 [0.5.2]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.5.2
 [0.5.3]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.5.3
+[0.6.0]: https://github.com/AusTrakka/austrakka2-cli/releases/tag/0.6.0